+++ src/resources/templates/index.ts
@@ -1,4 +1,12 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 export { Agents, type AgentCreateResponse, type AgentCreateParams } from './agents';
-export { Templates } from './templates';
+export {
+  Templates,
+  type TemplateCreateResponse,
+  type TemplateUpdateResponse,
+  type TemplateDeleteResponse,
+  type TemplateCreateParams,
+  type TemplateUpdateParams,
+  type TemplateDeleteParams,
+} from './templates';