+++ tests/api-resources/tools.test.ts
@@ -149,16 +149,4 @@ describe('resource tools', () => {
       tags: ['string'],
     });
   });
-
-  // Prism tests are disabled
-  test.skip('upsertBaseTools', async () => {
-    const responsePromise = client.tools.upsertBaseTools();
-    const rawResponse = await responsePromise.asResponse();
-    expect(rawResponse).toBeInstanceOf(Response);
-    const response = await responsePromise;
-    expect(response).not.toBeInstanceOf(Response);
-    const dataAndResponse = await responsePromise.withResponse();
-    expect(dataAndResponse.data).toBe(response);
-    expect(dataAndResponse.response).toBe(rawResponse);
-  });
 });